feat: Moved external item drag and drop functionality to component directory
