+++ projects/f-flow/src/f-external-item/domain/providers.ts
@@ -1,7 +1,7 @@
 import { EXTERNAL_ITEM_PREPARATION_PROVIDERS } from './external-item-preparation';
 import { EXTERNAL_ITEM_FINALIZE_PROVIDERS } from './external-item-finalize';
 
-export const EXTERNAL_ITEM_PROVIDERS = [
+export const F_EXTERNAL_ITEM_DRAG_AND_DROP_PROVIDERS = [
 
   ...EXTERNAL_ITEM_FINALIZE_PROVIDERS,
