+++ projects/f-flow/src/f-external-item/index.ts
@@ -1,3 +1,5 @@
+export * from './domain';
+
 export * from './f-external-item-base';
 
 export * from './is-external-item';
